Coords fix

+++ appartmenthostService/Models/Apartment.cs
@@ -27,9 +27,9 @@ namespace apartmenthostService.Models
 
         public string AdressTypes { get; set; }
 
-        public decimal? Latitude { get; set; }
+        public double? Latitude { get; set; }
 
-        public decimal? Longitude { get; set; }
+        public double? Longitude { get; set; }
 
         public string PlaceId { get; set; }
